Implemented Dark/Light Mode Switch
