+++ client/src/app/layout/Header.tsx
@@ -1,12 +1,18 @@
-import { AppBar, Toolbar, Typography } from "@mui/material";
+import { AppBar, Switch, Toolbar, Typography } from "@mui/material";
 
-export default function Header() {
+interface Props {
+    darkMode: boolean;
+    handleThemeChange: () => void;
+}
+
+export default function Header({darkMode, handleThemeChange}: Props) {
     return (
         <AppBar position="sticky" sx={{mb: 4}}>
             <Toolbar>
                 <Typography variant="h6">
                     Everything Gaming
                 </Typography>
+                <Switch checked={darkMode} onChange={handleThemeChange} />
             </Toolbar>
         </AppBar>
     )
